shipment,audit,blockchain implemented

+++ src/main/java/com/traceledger/module/production/exception/FactoryNotFoundException.java
@@ -10,7 +10,7 @@ public class FactoryNotFoundException extends RuntimeException {
 		super("Factory not found");
 	}
 
-	public FactoryNotFoundException(Integer id) {
+	public FactoryNotFoundException(Long id) {
 		super("Factory with id " + id+ " is not found");
 	}
